+++ domain/src/main/java/com/dida/domain/model/nav/community/CreateCommunityNft.kt
@@ -1,7 +1,7 @@
 package com.dida.domain.model.nav.community
 
 data class CreateCommunityNft(
-    val nftId: Int,
+    val nftId: Long,
     val nftImg: String,
     val nftName: String,
     val userImg: String,
