+++ model/src/main/java/com/dida/model/domain/Posts.kt
@@ -5,8 +5,8 @@ data class Posts(
 )
 
 data class PostItem(
-    val postId: Int,
-    val cardId: Int,
+    val postId: Long,
+    val cardId: Long,
     val userName: String,
     val userImgUrl: String,
     val title: String,
